+++ docs/tutorials/gallery_rest_api.py
@@ -16,7 +16,7 @@ from pathlib import Path
 
 sys.path.insert(0, str(Path.cwd().parent))
 
-import documentation_helpers
+import ragna._docs.documentation_helpers as documentation_helpers
 
 # %%
 # ## Step 1: Start the REST API
